pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "scribe"
version = "0.1.0"
description = "CLI transcription & course processor"
requires-python = ">=3.12"
dependencies = [
    "typer[all]>=0.9.0",
    "deepgram-sdk>=3.0.0",
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
scribe = "scribe.cli:app"

[tool.setuptools.packages.find]
where = ["."]
